+++ CHANGELOG.md
@@ -38,6 +38,12 @@ Bugfixes:
 
 - Strip byte order mark from all input strings, not just when loading files (#128) -- @Diggory
 
+## 0.8.2
+
+Bugfixes:
+
+- Throw an error when encountering duplicate column names in CSV headers (#136) -- @TomorrowMC
+
 ## 0.8.1
 
 Bugfixes: